Panels: remove minSize={20} for a reflow issue
When this property is set, a re-layout (force reflow) is performed by the browser even with a simple hovering of the separator.

Since we may have very large walls of text/images, we need to minimize relayouts, so for now, we set a min size on the contained scrollable container instead of preventing the resize.

See also this upstream issues: https://github.com/bvaughn/react-resizable-panels/issues/337