Flash/Backup Data
A new setting to export all the settings in localstorage and IndexedDB into
single 'flash' files for Big-AGI to reload.

This allows to quickly and easily migrate a full installation, including images,
from a v2-dev open installation to another.

This won't likely work across other branches, but it's meant to be forward-compatible.